move AddItems.js to pages folder, create an index for that folder
