+++ src/components/index.js
@@ -1,5 +1,4 @@
 import Modal from './Modal';
-import AddItems from './AddItems';
 
 // Then you can export all of the components from here
-export { Modal, AddItems };
+export { Modal };
