+++ src/pages/index.js
@@ -1,0 +1,3 @@
+import AddItems from './AddItems';
+
+export { AddItems };
